refactor(#121): Phase3 Step3.1 — consolidate process seam into analyzer/platform/process.py
- analyzer/launcher.py -> analyzer/platform/process.py (spawn_detached)
- pid-alive + win kernel32 logic moved from launch_dashboard.py into process.py
- launch_dashboard._spawn_server detach duplicate removed; now calls spawn_detached
- launch_dashboard / launch_archive / restart_dashboard share process.is_pid_alive
- new tests/test_platform_process.py pins the seam contract
- restart_dashboard.py:145 SIGKILL-escalation branch kept in place (fallback:
  not duplication, tightly coupled to terminate flow — documented residual site)

sys.platform == "win32" production occurrences: launch_dashboard.py and
launcher.py no longer carry OS branches; process seam owns them.
