+++ hooks/launch_dashboard.py
@@ -63,6 +63,7 @@ from typing import Optional
 # 直接 `python <abs path>/launch_dashboard.py` で起動されるため)。
 sys.path.insert(0, str(Path(__file__).resolve().parent.parent))
 from analyzer.server_registry import remove_server_json as _registry_remove_server_json
+from analyzer.platform.process import is_pid_alive, spawn_detached
 
 _DEFAULT_SERVER_JSON_PATH = Path.home() / ".claude" / "transcript-analyzer" / "server.json"
 SERVER_JSON_PATH = Path(os.environ.get("DASHBOARD_SERVER_JSON", str(_DEFAULT_SERVER_JSON_PATH)))
@@ -146,89 +147,6 @@ def _read_server_json(path: Path) -> Optional[dict]:
     return info
 
 
-# Issue #24: Windows では `os.kill(pid, 0)` が SystemError を上げる (Python issue 14480)。
-# OpenProcess + GetExitCodeProcess で alive 判定する。STILL_ACTIVE=259 は MSDN
-# `GetExitCodeProcess` ドキュメントで定数定義されている値 (確認済み)。
-_WIN_PROCESS_QUERY_LIMITED_INFORMATION = 0x1000
-_WIN_STILL_ACTIVE = 259
-
-# kernel32 の bound symbol キャッシュ。初回 `_win_kernel32()` 呼び出しで signature 設定 +
-# キャッシュ、以降の launch_dashboard 経路はキャッシュ参照のみで < 100ms budget を維持。
-_WIN_KERNEL32 = None
-
-
-def _win_kernel32():
-    """Windows kernel32 を取得し ctypes signatures を明示する (Issue #24 PR#31 codex P1)。
-
-    Win64 で `OpenProcess` の戻り値 HANDLE はポインタ幅 (64bit) だが、ctypes default の
-    `restype=c_int` (32bit signed) のままだと高位ビットが立った handle で truncate +
-    sign-extend が起き、`GetExitCodeProcess` / `CloseHandle` に誤った handle を渡す
-    可能性がある。`wintypes.HANDLE` で明示すれば正しく扱える。bound symbol を一度だけ
-    設定してキャッシュするので毎回 lookup が走らず launcher の起動 budget も安全。
-    """
-    global _WIN_KERNEL32  # pylint: disable=global-statement
-    if _WIN_KERNEL32 is not None:
-        return _WIN_KERNEL32
-    import ctypes  # pylint: disable=import-outside-toplevel
-    from ctypes import wintypes  # pylint: disable=import-outside-toplevel
-    k = ctypes.windll.kernel32  # type: ignore[attr-defined]
-    k.OpenProcess.restype = wintypes.HANDLE
-    k.OpenProcess.argtypes = [wintypes.DWORD, wintypes.BOOL, wintypes.DWORD]
-    k.GetExitCodeProcess.restype = wintypes.BOOL
-    k.GetExitCodeProcess.argtypes = [wintypes.HANDLE, ctypes.POINTER(wintypes.DWORD)]
-    k.CloseHandle.restype = wintypes.BOOL
-    k.CloseHandle.argtypes = [wintypes.HANDLE]
-    _WIN_KERNEL32 = k
-    return _WIN_KERNEL32
-
-
-def _is_pid_alive_posix(pid: int) -> bool:
-    """POSIX: `os.kill(pid, 0)` で存在確認。ESRCH → False、EPERM → True。"""
-    try:
-        os.kill(pid, 0)
-        return True
-    except ProcessLookupError:
-        return False
-    except PermissionError:
-        # 他ユーザの pid が偶然 collision したケース。「別プロセスが存在する」の意味で True
-        return True
-    except OSError:
-        return False
-
-
-def _is_pid_alive_windows(pid: int) -> bool:
-    """Windows: kernel32!OpenProcess + GetExitCodeProcess で alive 判定。
-
-    OpenProcess が NULL → 不在 or アクセス不可 → False。
-    GetExitCodeProcess の exit_code が STILL_ACTIVE (259) → 生存中 → True。
-    """
-    import ctypes  # pylint: disable=import-outside-toplevel
-    from ctypes import wintypes  # pylint: disable=import-outside-toplevel
-    kernel32 = _win_kernel32()
-    handle = kernel32.OpenProcess(
-        _WIN_PROCESS_QUERY_LIMITED_INFORMATION, False, pid
-    )
-    if not handle:
-        return False
-    try:
-        exit_code = wintypes.DWORD()
-        ok = kernel32.GetExitCodeProcess(handle, ctypes.byref(exit_code))
-        return bool(ok) and exit_code.value == _WIN_STILL_ACTIVE
-    finally:
-        kernel32.CloseHandle(handle)
-
-
-def _is_pid_alive(pid: int) -> bool:
-    """OS 別 dispatch。テストは sys.platform mock で各分岐を検証可能。"""
-    if sys.platform == "win32":
-        try:
-            return _is_pid_alive_windows(pid)
-        except OSError:
-            # ctypes 呼び出しが落ちた場合の保険。多重起動回避を優先して False
-            return False
-    return _is_pid_alive_posix(pid)
-
-
 def _healthz_ok(url: str, timeout: float = HEALTHZ_TIMEOUT_SECONDS) -> bool:
     """`{url}/healthz` が 200 を返すか。接続失敗 / timeout / 非 200 → False。"""
     try:
@@ -263,7 +181,7 @@ def _server_is_alive() -> bool:
     url = info.get("url")
     if not isinstance(pid, int) or not isinstance(url, str) or not url:
         return False
-    if not _is_pid_alive(pid):
+    if not is_pid_alive(pid):
         return False  # 副作用なし: 削除は _remove_stale_server_json() / main() で行う
     # pid alive 確定。healthz をリトライしてみるが、結果に関わらず True 返却
     # (race window 吸収 + 二重起動防止)。
@@ -282,7 +200,7 @@ def _remove_stale_server_json() -> bool:
     - 不在 / 壊れた JSON / 非 dict → 何もしない (False)。spawn 後の atomic replace に任せる
 
     Issue #24 PR#31 codex P2: 削除は `server_registry.remove_server_json` 経由で
-    `_file_lock` + compare-and-delete を経由する。事前 `_is_pid_alive` 判定後に
+    `_file_lock` + compare-and-delete を経由する。事前 `is_pid_alive` 判定後に
     他プロセスが新 server.json を上書きしても、pid 不一致で削除を skip するため
     安全。dashboard/server.py の write/remove と同じ lock を共有することで
     multi-hook race window で TOCTOU race が再発しない。
@@ -293,18 +211,9 @@ def _remove_stale_server_json() -> bool:
     pid = info.get("pid")
     if not isinstance(pid, int):
         return False
-    if _is_pid_alive(pid):
+    if is_pid_alive(pid):
         return False
     return _registry_remove_server_json(SERVER_JSON_PATH, expected_pid=pid)
-
-
-# Windows fork-and-detach 用の CreateProcess flags (subprocess.* は Win 限定属性)。
-# POSIX で AttributeError を避けるため getattr fallback でハードコード値を採用。
-# DETACHED_PROCESS=0x8 / CREATE_NEW_PROCESS_GROUP=0x200 (MSDN: process creation flags)
-_WIN_DETACHED_PROCESS = getattr(subprocess, "DETACHED_PROCESS", 0x00000008)
-_WIN_CREATE_NEW_PROCESS_GROUP = getattr(
-    subprocess, "CREATE_NEW_PROCESS_GROUP", 0x00000200
-)
 
 
 def _spawn_server() -> Optional[subprocess.Popen]:
@@ -313,37 +222,12 @@ def _spawn_server() -> Optional[subprocess.Popen]:
     Issue #34 Proposal 1: 戻り値で `Popen` を返すことで、main() 側で `proc is None`
     のとき poll を呼ばずに silent return できる (古い server.json の誤読防止)。
 
-    OS 別 detach 経路 (Issue #24):
-    - POSIX: `start_new_session=True` で親 PG/SID から切り離し、Claude Code 終了後も生存
-    - Windows: `creationflags=DETACHED_PROCESS|CREATE_NEW_PROCESS_GROUP` で同等の切り離し
-      (`start_new_session` は Win では no-op で親終了時に子も死ぬため不可)
-
-    共通:
-    - `stdin/stdout/stderr=DEVNULL`: 親 hook の pipe を引き継がない
-    - `close_fds=True`: 余計な fd を継承しない
+    OS 別 detach 経路は `analyzer.platform.process.spawn_detached` に集約済み
+    (Issue #121): POSIX は `start_new_session`、Windows は `creationflags`。
     """
     if not _SERVER_SCRIPT.exists():
         return None
-    kwargs: dict = {
-        "stdin": subprocess.DEVNULL,
-        "stdout": subprocess.DEVNULL,
-        "stderr": subprocess.DEVNULL,
-        "close_fds": True,
-    }
-    if sys.platform == "win32":
-        kwargs["creationflags"] = (
-            _WIN_DETACHED_PROCESS | _WIN_CREATE_NEW_PROCESS_GROUP
-        )
-    else:
-        kwargs["start_new_session"] = True
-    try:
-        return subprocess.Popen(  # pylint: disable=consider-using-with
-            [sys.executable, str(_SERVER_SCRIPT)],
-            **kwargs,
-        )
-    except OSError:
-        # PermissionError / FileNotFoundError / fork limit 等。silent fail。
-        return None
+    return spawn_detached([sys.executable, str(_SERVER_SCRIPT)])
 
 
 # ----------------------------------------------------------------------------
